feat: separate gosuki schema and buku compatibility view
- keep a virtual table `bookmarks` for compatibility with buku
- gosuki uses the internal `gskbookmarks` table
- add schema versioning for future upgrades
- add insert & update triggers to `bookmarks` view

Details: Previously there was only a `bookmarks` table which contained a
few extra columns compared to Buku. This table is renamed to
`gskbookmarks` which will be the native gosuki table schema.

Instead we provide a `bookmarks` view into `gskbookmarks` with INSERT
and UPDATE trigger (INSTEAD OF) that allow Buku programs to use Gosuki
DB as a buku database while still benifiting from gosuki specific
features that will eventually be added to the schema.

Added also schema versioning that will be tracked in the table
creatively named `schema_version`.